Create Gunika
Having an experience of 5 years in total and in few different fields has only helped me learn more about different aspects of working in different kind of fields(counselling, research and entrepreneurship) and also learn exuberantly about market and primary and secondary research and finding a new born interest in the same. Being a mass communication student I have always looked forward to connect to masses in various manner.  There are numerous reasons for not finding me on social media as much people always expected me to because of me being mostly a patient of being shy and which is why I chose writing as my permanent career choice and I enthusiastically call myself a writer now.